% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Create a shiny server from an object and its function list}
\usage{
app_server(id = "r6interface", obj)
}
\arguments{
\item{obj}{an R6 object}

\item{obj_funs}{}
}
\value{
a shiny module server that can be fed into a shiny::shinyApp call or
included as server logic in another app - possibly in a shiny.router
application page
}
\description{
Create a shiny server from an object and its function list
}
