% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Create a shiny ui shell for object and its function list}
\usage{
app_ui(id = "r6interface", obj_name)
}
\arguments{
\item{obj_name}{The name of the object that is being interfaced with}
}
\value{
html ui for a Shiny app
}
\description{
Create a shiny ui shell for object and its function list
}
