% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{Deck}
\alias{Deck}
\title{A deck of cards}
\description{
A deck of cards class compatible with Interfacing
}
\section{Fields}{

\describe{
 \item{max_state}{The baseline state of Deck that it may return to when `reshuffle(0L)`d.}
 \item{state}{The current state of the Deck.}
 \item{hands}{The list of hands currently drawn from the Deck.}
}
}

\section{Methods}{

\code{$new()} Creates a new Deck object.
\code{$print()} Returns a list containing the Deck state and list of hands.
\code{$reshuffle(hand_index = 0L)} If hand_index is 0, returns Deck to its max_state. Otherwise, reshuffles the hand at `hand_index` into the Deck `state`.
\code{$draw(ncards = 5)} Draws a hand of `ncards` from `state` and adds them to `hands`.
}

