% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{Dice}
\alias{Dice}
\title{A set of dice}
\description{
A set of dice class compatible with Interfacing
}
\section{Fields}{

\describe{
 \item{state}{The current state of all pools.}
 \item{pools}{The list of dice pools currently formed.}
}
}

\section{Methods}{

\code{$new()} Creates a new Dice object.
\code{$print()} Returns a list containing the Deck state and list of hands.
\code{$add_pool()} Creates a named pool with the specified dice
}

