% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{makeShinyApp}
\alias{makeShinyApp}
\title{Create a shiny app from an R6 object}
\usage{
makeShinyApp(obj)
}
\arguments{
\item{obj}{an R6 object}
}
\value{
returns a shiny app that can be launched via shiny::runApp
}
\description{
Create a shiny app from an R6 object
}
