% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Create a shiny app from an R6 object}
\usage{
runShinyApp(obj)
}
\arguments{
\item{obj}{an R6 object}
}
\value{
makes and runs the shiny app created by makeShinyApp
}
\description{
Create a shiny app from an R6 object
}
